+++ iti-react/Changelog.md
@@ -842,3 +842,4 @@ const myTab: Tab = ['tabName', 'Tab Label', { className: 'my-tab' }]
 - `PersonNameInput`:  
     - \*\*\* `disabled` is no longer allowed in `inputAttributes`. Use the `enabledInputs` prop instead.
     - Convert to `useValidation`
+- LOTS of small changes to fix eslint errors